requirements.txt
airtable-python-wrapper==0.11.2
appdynamics==4.5.0.0
appdynamics-bindeps-osx-x64==7.0
appdynamics-proxysupport-osx-x64==1.8.0.51.1
asn1crypto==0.24.0
cachetools==2.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
dill==0.2.7.1
Django==1.9
django-colorfield==0.1.14
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
django-gcloud-storage==0.2.1
django-hstore==1.4.2
django-jet==1.0.7
django-modeladmin-reorder==0.3.1
django-nested-inline==0.3.7
django-reset-migrations==0.3.1
django-rest-framework-mongoengine==3.3.1
django-rest-swagger==2.1.2
django-storages==1.6.6
django-tinymce==2.7.0
djangorestframework==3.6.2
djangorestframework-gis==0.13
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
drf-openapi==1.3.0
drfdocs==0.0.11
enum34==1.1.6
future==0.16.0
futures==3.2.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.4
google-auth==1.5.0
google-cloud==0.33.1
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.2
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.13.0
gunicorn==19.8.1
httplib2==0.11.3
idna==2.6
ipaddress==1.0.22
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauth2client==3.0.0
openapi-codec==1.3.2
Pillow==5.1.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.2.post1
psutil==5.4.5
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyJWT==1.6.4
python-dotenv==0.8.2
python-http-client==3.1.0
pytz==2017.3
requests==2.18.4
rsa==3.4.2
sendgrid==5.6.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.14.1
whitenoise==3.3.1
